requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
python-dotenv
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
email-validator